requirements.txt
playwright>=1.40.0
robotframework>=6.0.0
